+++ libs/gui/include/mrpt/gui/WxSubsystem.h
@@ -405,7 +405,7 @@ class C3DWindowDialog : public wxFrame
 	void addTextMessage(
 		const double x_frac, const double y_frac, const std::string& text,
 		const mrpt::img::TColorf& color, const std::string& font_name,
-		const double font_size, const mrpt::opengl::TOpenGLFontStyle font_style,
+		const float font_size, const mrpt::opengl::TOpenGLFontStyle font_style,
 		const size_t unique_index, const double font_spacing,
 		const double font_kerning, const bool has_shadow,
 		const mrpt::img::TColorf& shadow_color);
