+++ libs/opengl/src/CEllipsoidRangeBearing2D.cpp
@@ -17,8 +17,7 @@ using namespace mrpt::opengl;
 using namespace mrpt::math;
 using namespace std;
 
-IMPLEMENTS_SERIALIZABLE(
-	CEllipsoidRangeBearing2D, CRenderizable, mrpt::opengl)
+IMPLEMENTS_SERIALIZABLE(CEllipsoidRangeBearing2D, CRenderizable, mrpt::opengl)
 
 /*---------------------------------------------------------------
 							transformFromParameterSpace
@@ -34,10 +33,10 @@ void CEllipsoidRangeBearing2D::transformFromParameterSpace(
 	out_pts.resize(N);
 	for (size_t i = 0; i < N; i++)
 	{
-		const double range = in_pts[i][0];
-		const double bearing = in_pts[i][1];
-		out_pts[i][0] = range * cos(bearing);
-		out_pts[i][1] = range * sin(bearing);
+		const float range = in_pts[i][0];
+		const float bearing = in_pts[i][1];
+		out_pts[i][0] = range * cosf(bearing);
+		out_pts[i][1] = range * sinf(bearing);
 	}
 
 	MRPT_END
